Fixed Project structure
Done :

- Split widget and section files into separate files for each widget
- Created a pages folder for all main pages
- Created a logic folder for all backend logic
- Added a provider folder for future improvements

TODO :
- Providers ( PLEASE )
