+++ lib/Widgets/components/Shimmer.dart
@@ -1,149 +1,5 @@
+
 import 'package:flutter/material.dart';
-import 'package:lottie/lottie.dart';
-
-class DayTile extends StatelessWidget {
-  const DayTile({
-    super.key,
-    required this.nextDays,
-    required this.index,
-  });
-  final index;
-  final List nextDays;
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: [
-        SizedBox(
-            width: 55,
-            child:
-                StyledText(nextDays[index].day, 'grillmaster', 24, 'normal')),
-        Column(
-          crossAxisAlignment: CrossAxisAlignment.center,
-          children: [
-            Stack(
-              alignment: Alignment.bottomCenter,
-              children: [
-                Image.network(nextDays[index].icon),
-                Positioned(
-                    bottom: 10,
-                    child: StyledText(
-                        nextDays[index].condition, 'grillmaster', 10, 'normal'))
-              ],
-            )
-          ],
-        ),
-        SizedBox(
-          width: 123,
-          child: Row(children: [
-            StyledText(
-                '${nextDays[index].maxTemp}°', 'PublicSans', 45, 'normal'),
-            const StyledText('/', 'PublicSans', 25, 'normal'),
-            StyledText(
-                '${nextDays[index].minTemp}°', 'PublicSans', 20, 'normal'),
-          ]),
-        )
-      ],
-    );
-  }
-}
-
-class GradientText extends StatelessWidget {
-  const GradientText({
-    super.key,
-    required this.text,
-    this.style,
-    required this.gradient,
-  });
-  final String text;
-  final TextStyle? style;
-  final Gradient gradient;
-  @override
-  Widget build(BuildContext context) {
-    return ShaderMask(
-      blendMode: BlendMode.srcIn,
-      shaderCallback: (bounds) => gradient.createShader(
-        Rect.fromLTWH(0, 0, bounds.width, bounds.height),
-      ),
-      child: Text(text, style: style),
-    );
-  }
-}
-
-class InfoBox extends StatelessWidget {
-  const InfoBox({
-    super.key,
-    required this.iconPath,
-    required this.infoTitle,
-    required this.infoValue,
-  });
-  final infoTitle;
-  final iconPath;
-  final infoValue;
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: 155,
-      child: Row(
-        children: [
-          Lottie.asset(iconPath, height: 50, width: 50),
-          const SizedBox(
-            width: 10,
-          ),
-          Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              StyledText(infoTitle, 'grillmaster', 14, 'normal'),
-              StyledText(infoValue, 'PublicSans', 20, 'bold'),
-            ],
-          )
-        ],
-      ),
-    );
-  }
-}
-
-class NextDaysList extends StatelessWidget {
-  final nextDays;
-
-  const NextDaysList({super.key, required this.nextDays});
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: [
-        for (var i = 0; i < 7; i++) ...[
-          DayTile(
-            index: i,
-            nextDays: nextDays,
-          ),
-          const Divider(),
-        ],
-      ],
-    );
-  }
-}
-
-class StyledText extends StatelessWidget {
-  const StyledText(this.text, this.family, this.size, this.weight, {super.key});
-  final String text, family;
-  final double size;
-  final String weight;
-  @override
-  Widget build(BuildContext context) {
-    return Text(
-      text,
-      style: TextStyle(
-          color: Colors.white,
-          fontFamily: family,
-          fontSize: size,
-          fontWeight: weight == 'bold' ? FontWeight.bold : FontWeight.normal),
-    );
-  }
-}
 
 class Shimmer extends StatefulWidget {
   static ShimmerState? of(BuildContext context) {
