+++ lib/models/CurrentWeatherModel.dart
@@ -49,53 +49,3 @@ class CurrentWeather {
     );
   }
 }
-
-//build second class for forecaast part
-class ForecastWeather {
-  final List<dynamic> day;
-  final List<dynamic> condition;
-  final List<dynamic> minTemp;
-  final List<dynamic> maxTemp;
-  const ForecastWeather({
-    required this.day,
-    required this.condition,
-    required this.minTemp,
-    required this.maxTemp,
-  });
-
-  factory ForecastWeather.fromJson(Map<String, dynamic> json) {
-    return ForecastWeather(
-      day: json['daily']['time'],
-      condition: json['daily']['weather_code'],
-      minTemp: json['daily']['temperature_2m_min'],
-      maxTemp: json['daily']['temperature_2m_max'],
-    );
-  }
-}
-
-class NextDays {
-  final String day;
-  final String icon;
-  final String condition;
-  final String minTemp;
-  final String maxTemp;
-  NextDays(
-      {required this.day,
-      required this.icon,
-      required this.condition,
-      required this.minTemp,
-      required this.maxTemp});
-}
-
-class SearchLocation {
-  final String city;
-  final String country;
-  final String lat;
-  final String lng;
-  SearchLocation({
-    required this.city,
-    required this.country,
-    required this.lat,
-    required this.lng,
-  });
-}
