+++ lib/models/NextDaysmModel.dart
@@ -1,0 +1,13 @@
+class NextDays {
+  final String day;
+  final String icon;
+  final String condition;
+  final String minTemp;
+  final String maxTemp;
+  NextDays(
+      {required this.day,
+      required this.icon,
+      required this.condition,
+      required this.minTemp,
+      required this.maxTemp});
+}
