+++ lib/models/SearchLocationModel.dart
@@ -1,0 +1,12 @@
+class SearchLocation {
+  final String city;
+  final String country;
+  final String lat;
+  final String lng;
+  SearchLocation({
+    required this.city,
+    required this.country,
+    required this.lat,
+    required this.lng,
+  });
+}
